+++ docs/changelog.lisp
@@ -15,7 +15,7 @@
          "
 ## Added
 
-* Variable 40ANTS-LOGGING:*ON-CHANGE-HOOKS* to run custom callbacks after logging configuration changes such as SETUP-FOR-BACKEND, SETUP-FOR-CLI, SETUP-FOR-REPL, and REMOVE-REPL-APPENDER.
+* Variable 40ANTS-LOGGING:*ON-CHANGE-HOOKS* to run custom callbacks after logging configuration changes such as 40ANTS-LOGGING:SETUP-FOR-BACKEND, 40ANTS-LOGGING:SETUP-FOR-CLI, 40ANTS-LOGGING:SETUP-FOR-REPL, and 40ANTS-LOGGING:REMOVE-REPL-APPENDER.
 ")
   (0.4.0 2025-10-31
          "